requirements.txt
openai-whisper
torch
torchvision
torchaudio
soundfile
pandas
scikit-learn
numpy
